requirements.txt
# FastAPI Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# AI/ML - Original
google-generativeai==0.8.3

# LlamaIndex - RAG Framework with Conversation Memory
llama-index-core>=0.10.0
llama-index-vector-stores-postgres>=0.1.0
llama-index-embeddings-gemini>=0.1.0
llama-index-llms-gemini>=0.1.0
llama-index-storage-chat-store-postgres>=0.1.0

# Database
psycopg2-binary==2.9.10
pgvector==0.3.6
sqlalchemy==2.0.36

# Document Processing
pypdf==5.1.0
langchain-text-splitters==0.3.4

# Configuration & Environment
python-dotenv==1.0.1
pydantic==2.10.3

# HTTP Client
requests>=2.31.0
httpx>=0.27.0

# Authentication
python-jose[cryptography]>=3.3.0

# Supabase Storage - PHASE 2
supabase>=2.4.0

# Vercel/Serverless adapter
mangum>=0.17.0

# Optional: Development tools (uncomment for development)
# pytest==8.3.3
# pytest-asyncio==0.24.0
# black==24.8.0
# ruff==0.7.0
